♻️ refactor: initData config 분리 및 test 일부 수정
